rewrite get_loss and associated tests additional linting in lib
