+++ tests/test_data.py
@@ -1,6 +1,6 @@
-from pathlib import Path
 import tempfile
 import unittest
+from pathlib import Path
 
 import monai
 from monai.data import DataLoader
