+++ trainlib/model.py
@@ -1,5 +1,6 @@
-from monai.networks.nets import UNet
 import munch
+from monai.networks.nets import UNet
+
 
 def get_model(config: munch.Munch):
     """Create a standard UNet"""
