+++ trainlib/transforms.py
@@ -2,9 +2,9 @@ import inspect
 from typing import Callable, List
 
 import monai
+import munch
 from monai.transforms import Compose
 from monai.utils.enums import CommonKeys
-import munch
 
 from trainlib.utils import import_patched
